# Fix #82: Optimize funcqc startup time by 70-80% for read-only commands